Update team_test to be singular, and to match team instance variables
