+++ lib/game_team.rb
@@ -33,5 +33,4 @@ class GameTeam
     @giveaways = details[:giveaways].to_i
     @takeaways = details[:takeaways].to_i
   end
-
 end
